requirements.txt
pycryptodome==3.21.0
pycoingecko==3.2.0